Fix chain redirection porblem : eg : cat << eof > a.txt
